refactor(web): remove the xterm/PTY renderer; one live view everywhere
Commits to the unified renderer: the agent and paired-shell panes now use
the capture-snapshot LiveTerminalView on every device (the TUI's live-mode
model), and the xterm.js PTY-relay rendering is gone. No fallback, no
pointer-type dispatch.

Removed:
- XtermTerminalView and the inner xterm PairedTerminal; TerminalView and
  PairedShellPane render LiveTerminalView unconditionally.
- The useTerminal hook (xterm + PTY relay) and its tests; the reconnect
  backoff schedule moved to lib/wsBackoff.ts (kept under test).
- The @xterm/* dependencies.
- Desktop-xterm-only test specs whose behaviors no longer exist with this
  renderer: wheel-scroll (SGR escapes), pinch-zoom-desktop (Ctrl+wheel
  zoom), terminal-ime (xterm IME), init-resize-storm, paired-hidden-resize,
  ws-reconnect, terminal-copy-select, and the desktop branches of
  backspace/focus/toolbar specs. "terminal loaded" waits repointed from
  `.xterm` to `[data-live-terminal]`.

The touch-only chrome (soft-keyboard toolbar + FAB) is gated off
fine-pointer devices so the desktop live view is clean.

Deliberate consequence: desktop loses xterm-era features (mouse-wheel
scroll forwarded to the agent, Ctrl+wheel font zoom, xterm IME, instant
PTY echo); input now round-trips via send-keys + capture polling like
mobile. Accepted to collapse two web pipelines into one.

The server PTY WS routes (/ws, /terminal/ws, /container-terminal/ws) are
now unused by the frontend; removing that relay is a separate follow-up.

Verified: cargo build --features serve, web tsc/eslint/oxfmt, vitest 1814,
mocked Playwright 291, live golden-path/size-owner/no-flutter green.
