+++ web/package.json
@@ -24,10 +24,6 @@
     "@dnd-kit/utilities": "^3.2.2",
     "@pierre/diffs": "^1.2.5",
     "@tailwindcss/vite": "^4.2.2",
-    "@xterm/addon-fit": "^0.11.0",
-    "@xterm/addon-web-links": "^0.12.0",
-    "@xterm/addon-webgl": "^0.19.0",
-    "@xterm/xterm": "^6.0.0",
     "clsx": "^2.1.1",
     "cmdk": "^1.1.1",
     "lucide-react": "^1.14.0",
